# Conflicts:
#   routes/web.php

+++ app/Http/Controllers/EventController.php
@@ -9,6 +9,7 @@ use App\Models\Tag;
 
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
@@ -72,6 +73,10 @@ class EventController extends Controller
 
         if (!Auth::check()) return redirect('/login');
 
+        if(Session::has('orig_user')) {
+            return redirect()->to('/')->withErrors('You are not authorized to create an event in user view.');
+        }
+
         $this->validator($request->all())->validate();
         $user = Auth::user();
         $event = Event::create([
@@ -124,6 +129,9 @@ class EventController extends Controller
     public function editEvent(Request $request)
     {
         if (!Auth::check()) return redirect('/login');
+        if(Session::has('orig_user')) {
+            return redirect()->to('/')->withErrors('You are not authorized to edit an event in user view.');
+        }
         $this->validator($request->all())->validate();
         $user = Auth::user();
         $event = Event::create([
@@ -168,7 +176,9 @@ class EventController extends Controller
     public function deleteEvent($id)
     {
         if (!Auth::check()) return redirect('/login');
-
+        if(Session::has('orig_user')) {
+            return redirect()->to('/')->withErrors('You are not authorized to delete an event in user view.');
+        }
         $event = Event::find($id);
        // $event->user()->detach();
         $event->tags()->detach();
